feat(torch-frontend): Adds a new file for layer functions for functions like multi_head_attention and lstm
